拆分推送
